put executor manager in a separate file
